+++ NEWS.md
@@ -10,6 +10,7 @@
 - "geodist" package is now a hard dependency ("Imports" not "Suggests").
 - Add @leoniedu as package contributor (#282).
 - Fix updating of edge times in `insert_one_edge`; #278 thanks to @leoniedu
+- Fix `add_nodes_to_graph()` when graph has duplicated edges; #285 thanks to @leoniedu
 
 ---